refactor: scan page
