+++ src/lib/components/molecules/Modal.svelte
@@ -1,0 +1,31 @@
+<script lang="ts">
+	import { fly } from 'svelte/transition';
+    export let isModalOpen: boolean
+    export let closeCb: () => void
+    const closeModal = () => {
+		isModalOpen = false;
+		closeCb();
+	}
+</script>
+
+<ion-modal
+	is-open={isModalOpen}
+	backdrop-dismiss={false}
+	initial-breakpoint={0.6}
+	backdrop-breakpoint={0.8}
+	transition:fly
+	class="visible"
+>
+	<hr />
+	<ion-content class="ion-padding visible">
+		<ion-toolbar>
+			<ion-title>Results</ion-title>
+			<ion-buttons slot="end">
+				<ion-button color="danger" on:click={closeModal} on:keydown={closeModal} aria-hidden>Close</ion-button>
+			</ion-buttons>
+		</ion-toolbar>
+			<div class="mt-4 flex flex-col gap-2">
+                <slot/>
+			</div>
+	</ion-content>
+</ion-modal>
